refactor(ui): organize test suite page hook
